requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.5
django-cors-headers==3.2.1
django-grappelli==2.14.1
django-ical==1.7.0
django-oauth-toolkit==1.3.2
django-recurrence==1.10.2
djangorestframework==3.11.0
djangorestframework-camel-case==1.1.2
drf-yasg==1.17.1
icalendar==4.0.5
idna==2.9
inflection==0.4.0
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.9
